MXS-4761 The service must be spefified for the Urat service
Currently the prepare stage rewires the original service. That's
non-optimal as it means that new sessions then will use the Urat
service as their target _before_ the Urat service has been started,
which means that they later anyway need to be restarted.

Thus, it is better if the rewiring is done when the Urat service
is started. But that can only be done if the Urat service knows
which service it was installed for.