MXS-4761 Collect info to UratResult in UratBackend
There will be no flow-control, i.e. there may be multiple calls
to to UratBackend::write() before UratBackend::finish_result()
is called. Thus, there will need to be a deque<UratResult>
in UratBackend so that a finish_result()-call can be matched
to the right write()-call. First step is to not maintain any
state information in UratBackend itself; the second step will
be to introduce the deque.