MXS-4761 Remove other-to-other dependency
Earlier a "round" was defined as the same request having been
sent to and all responses received from the main server and all
other servers. That meant that in the case one "other" would be
very slow, then we would have to wait for that before any results
could be logged. Now that dependency has been removed, and every
"other" is only dependent on "main".