MXS-4761 Execute EXPLAIN when needed
When a ComparatorOtherResult is received by ComparatorSession,
a decision is made whether the execution needs to be EXPLAINed.
If so, a CompartorExplainResult is pushed to the deque of results
and an "EXPLAIN ..." statement is sent to the backend. When the
result is ready, it will be deliverd to ComparatorSession, which
will log the result.

The lifetime is managed by ComparatorExplainResult shared pointing
to ComparatorOtherResult, which is shared pointing to
ComparatorMainResult.