MXS-4761 Move json generation to stats-hierarchy
Since there's only going to be one "main" there is now depth
in the hierarchy that strictly speaking is not necessary.
However, the aim is that the json-output is readable, without
it being rendered by some UI. With the extra dept, the "main"
and other "output" (of which there probably is going to be
just one) are horizontically aligned.