MXS-4761 Post-pone EXPLAINs if necessary
An EXPLAIN cannot be executed immediately if a multi-part
request is in process, but has to be postponed for later.
Once there is a slot, all pending EXPLAINs can be executed
in one go.

Basically there is a possiblity that the session is not in
exactly the same state as it was in when the statement being
EXPLAINed originally was executed, but preventing that would
require stopping the request stream until it is know whether
the most recent request requires EXPLAINing.