MXS-4761 Remove unnecessary sub-states
It's only the suspending of sessions that must be handled
asynchronously, the rest will either succeed or fail. Thus,
there is no reason to treat the different steps as separate
states. Doing it like this means also that it will be easier
to share functionality betwen setup and tear-down (to be
implemented).