MXS-4761 Remove creation of monitor and fetch Gtids manually
The monitor was basically only needed for the Gtids, so to remove
the need for creating it the Gtids are now fetched manually. In
addition, the server status that the monitor would bring are not
really needed as a failure to connect the _other_ must not prevent
the the session from being created.