MXS-4761 Require that the comparison kind is specified
The comparison can be made to a primary or to replica. In
the former case, the replication must be stopped at the start
and preferably reset and started at the end but in the latter
case the replication must not be stopped at the start and hence,
obviously, neither reset nor started at the end.