MXS-4761 Execute EXPLAIN for main as well
Currently it is basically possible to configure things so that
only main is EXPLAINed, but that will be changed so that either
other or other and main are explained, but not main alone.

The arrangement is similar to that of ordinary main/other
requests. The lifetime of ComparatorExplainMainResult is
controlled by ComparatorExplainOtherResult that contains a
shared_ptr to the former. When the former is ready it notifies
that it is ready and when the latter is ready, it checks whether
the former already is.

Right now there is a general life-time issue; if other is faster
than main, the main result will actually not be waited for but
other result simply discarded. To be addressed in the future.