MXS-4761 Handle lifetimes properly
To ensure that result objects related to main and other stay
alive as long as necessary, they both need have an std::shared_ptr
to each other. That cycle is explicity broken by other, when it
removes itself from the dependency set of main.
An implication of this is that other cannot add itself to main in
the constructor as it is not possible to call shared_from_this()
before there exists a std::shared_ptr to this.

Further, EXPLAINs sent to main need to be sent after the relevat
clientReply() calls have been made all the way to the top as
otherwise there will be re-entrancy problems. That is now handled
using an lcall(), which will be replaced with a
MXS_SESSION::delay_routing() call.