MXS-4761 Never write from clientReply()
In the case of 'other' it happened to work to write to
the backend from within clientReply(), but that should
not be done and should there be an assert for preventing
that, it would have fired.

There are currently several calls made from the router
session to the backend object to set it up. Now that the
router session is provided, they can be clean up (in another
commit).