MXS-5004: Don't sort already sorted data
Reading, sorting and writing the data is not needed if it's already been
done. On debug builds, the difference in speed is very noticeable but
even in release mode builds this would be wasteful.

The new .tx file contains the transformation report which could be used
as the generic capture data report. This way the report on a capture is
only done once and only as a post-processing step. Right now the file
contains a placeholder value that is just used to mark that the
transformation was done.