MXS-4761 Ensure that results are freed
When waiting for responses, the 'main' result object and 'other'
result object will both maintain a shared_ptr to each other. That
circular dependency is broken when a response has been received
from both.

It seems that there currently is something that rarely causes
this not to happen, which subsequently would cause both of them
to leak. To ensure that all objects are freed, the circular
dependency is now, if needed, explicitly broken in the backend
destructor and a warning is logged.