MXS-5004: Add conversion from QueryEvents to RepEvents
This makes it possible to compare a single replay to the capture. While
this is very much an apples-to-oranges comparison, it at least serves to
verify that the metrics are roughly similar.

Adding more options to the --mode option would work but it would simply
be another way to pass what are essentially subcommands. As such, with
minor changes it can be converted into an actual command.

The convert command also makes it possible to convert from one input
format to another.