MXS-5004: Track returned and examined rows in replay
Assuming that the query replays identically each time, a change in the
number of rows examined by the optimizer is likely performance related
whereas a change in the number of returned rows is related to the
application logic and could indicate incorrectness. By storing both, the
two can be compared at the event level. The rows examined should be a
more stable per-event metric than the query latency. If the number of
rows returned by two replays matches but the number of rows examined
does not, it's an indication that a potentially worse query plan was
chosen.

The --no-row-counts option disables the extra SHOW STATUS query that the
player does and sets the rows_read value to zero. This makes it possible
to avoid the overhead if it is deemed to be too much.

The number of page reads done by InnoDB could also be tracked if they
were added to the SHOW STATUS output (MDEV-33724). Without
MDEV-MDEV-33724, tracking of them is only possible via ANALYZE
FORMAT=JSON which makes it much more difficult to implement.