Add command for dumping the canonicals
This is easier than having to run a replay to get them to be
generated. If capture sanitization or other sorts of value replacement
is to be done, this information needs to be available so that the
correct canonical IDs can be passed to whatever sanitization is done.

It'll also be a convenient tool to estimate the size of the working set
of SQL and how many statements there are.