Minor help generation improvements
Calculate the indents based on the lengths of the options and
commands. This way longer commands automatically adjust the indent
width.

Store the commands in the order they are declared. This makes it easier
to have the more common commands at the top.

The --output option was also missing the help text.