Use MXB_S logging macros for output
This automatically adds timestamps to all messages which makes it easier
to detect how long things took. It also makes it easy to save the output
to a different file if that's ever needed.

Also include capdefs.hh in all headers as the first header. This makes
sure that the module name is consistently used in the log output.

The syslog output is also silenced so that it only prints to stdout.