Use DuckDB for processing the CSV file
This is a convenient middle ground between inserting the data into a
real database and having to process it manually. Scaling the data to
sizes that don't fit into memory should also work out of the box.

Added a requirements.txt file so that the dependencies can be installed
by pip.