MXS-4761 Turn DiffExplainResult into a template
With this change, the class hierarchy rearrangement is done.
Now the two branches are 'main' and 'other', which then are split
further into separate classes for ordinary and explain results.

Getting rid of the dynamic_casts. which was the primary reason for
this excersize, still require some common functionality to be moved
from DiffOrdinaryOtherResult and DiffExplainOtherResult to the base
DiffOtherResult. Likewise in the main hierarchy.