Add comparison to postprocess.py
The compare compares two processed replays and calculates the difference
in their statistics. The returned list can then be sorted based on the
differences.

Also added in code that handles the argument parsing when used as a
command line script. This makes it possible to also include the file as
a python module which makes it easier to use it in a Jupyter notebook
for visualization.