MXS-4761 Introduce Diff[Main|Other]Result templates
Now there are two branches in the DiffResult hierarchy; one
related to results from 'main' and one related to results from
'other'. With these changes, it is now possible to remove the
dynamic_casts.

In the beginning there was void, but then there was the following:

         DiffBackend<>------------------------DiffResult
              |                                   |
     +--------+--------+                 +--------+--------+
     |                 |                 |                 |
DiffMainBackend DiffOtherBackend   DiffMainResult   DiffOtherResult

Some things are omitted here. For instance, DiffMainResult and
DiffOtherResult refer to each other using shared_ptr:s so there is
a circular dependency, which is explicitly broken when when we have
gotten a response from both main and other backend.

Then EXPLAIN results were added.

           DiffResult
               |
      +--------+--------+--------------------+
      |                 |                    |
DiffMainResult   DiffOtherResult      DiffExplainResult
                                             |
                                 +-----------+-----------+
                                 |                       |
                       DiffExplainMainResult   DiffExplainOtherResult

DiffOtherResult and DiffExplainOtherResult had to some extent similar
behaviour, but that was duplicated.

This worked fine, until the reality imposed the requirement that in
certain situations something needs to be done to any pending results
when a DiffOtherBackend is destructed.

In the destructor of DiffOtherBackend, the DiffResults where cast
into DiffOtherResult and some actions were taken. Otherwise fine,
except for the fact that some of those DiffResults were not
DiffOtherResults, but DiffExplainOtherResults and the conseqeuences
were dire. That's why the dynamic_casts were taken into use.

With the changes that have been done in this and earlier commits,
the hierarchy has basically (all implementation detail intermediary
templates have been excluded) been turned into the following.

                             DiffResult
                                 |
           +---------------------+------------------+
           |                     |                  |
DiffOrdinaryMainResult DiffExplainMainResult DiffOtherResult
                                                    |
                                        +-----------+-----------+
                                        |                       |
                            DiffOrdinaryOtherResult   DiffExplainOtherResult

Thus, it is now safe to cast in the destructor of DiffOtherBackend
all DiffResults into DiffOtherResults and do whatever is necessary.