Add interactive Jupyter notebook
The notebook displays some of the results of the processed JSON
summaries in an interactive way. Most of the code is located in the
notebook_helper.py file to make it easier to modify the code in external
editors and to review any changes: the text form of .ipynb file is JSON
and thus much harder to review.

The notebook can be used with the output of both the wcar and diff
modules. Currently, only the wcar output is guaranteed to work.