Install the reporting code
The Python scripts are installed as a part of the enterprise package but
they aren't installed into /usr/bin as they require python modules to be
installed and those aren't needed for the capture part of the feature.

Renamed the example file paths in the notebook to correspond with the
to-be-added README file.