Add README and helper scripts for visualization
The README is also installed into the package as it's currently not
visible anywhere else.

The helper scripts maxvisualize and maxpostprocess are a small layer of
bash script on top of the python code. These are installed into /usr/bin
but the code itself lives in /usr/share/maxscale/wcar.