MXS-4761 Drop "total" response distribution
mxs::ResponseDistribution will be replaced with an implementation
that first collects samples and then decides what kind of bins to use.
Doing that for the "total" case would then have to be done across all
sessions as requests happens and that seems a bit costly and/or messy.
So, at least for for the time being the "total" histogram is dropped.