Fix QPS and histogram calculation
The QPS calculation can be simplified to just a simple plot of x and y
values. Making it behave like a histogram was more complex than it
needed to be and the same could be achieved by just injecting a dummy
value at the end of the list. Similarly, the plotting of the QPS can be
made more tolerant of malformed inputs by just making sure that the
lengths are what they are expected to be.

The histogram calculation also didn't work if there was only one value
as the divisor ended up being zero.

Removed the output redirection from the maxvisualize script to make
figuring out errors easier. The output also contains the URL for the
visualization that is needed if it's run inside of a docker container.