Add SimTime::real_now()
This is needed by the RepEvents so that the time that is measured is the
real execution time instead of the simulated time. Otherwise, with
--speed=0 the simulated time is always the same and the duration
measurement becomes useless.

Instead of having the real_now() function behave like now() in that it
requires a call to tick() before the value is updated, calling
mxb::Clock::now() always will eliminate the need for the CAS that
updates the stored time value. This should be a lighter operation for
the replay compared to the old tick().now() call chain where the work
done by the CAS was always wasted.