Add `show` command to maxplayer
The `show` command shows the SQL of one or more query events from the
capture. This is a useful tool for figuring out which queries were
responsible for the outliers and whether they are the same for two
subsequent replays.

Renamed the `list-queries` command to `canonicals`. This describes it
better as the output in all examples is always `canonicals.csv`.