MXS-4761 Add dynamic creation of histogram bins
mxs::ResponseDistribution is too coarse to be useful. There
could be a significant difference, yet there would appear to
be none. Further, a histogram should have bins of equal width.

The problem in the case of Diff, is that you do not know up
front what the bins should be. For a "perfect" result you
should collect all values and then at the end calculate what
the bins should. In practice, that is not possible without
writing the results to disk and that is not an option in the
case of Diff. The chosen solution is to collect a certain
number of samples, calculate the bins based on those values
and then use them.

The samples are collected in DiffRouter::add_sample_for(...)
and that will be a hotspot while the samples are collected.
However, eventually, when all samples have been collected,
the mutex will be locked just once when a copy of a
shared_ptr is made.