MXS-5043 Add initial QPS calculation
The QPS is calculated separately for 'main' and 'other'. In practice
there will be very little difference between 'main' and 'other', unless
there is some serious problems with either and e.g. SELECTS are dropped
to bring 'other' back in line.

Currently the QPS is collected from the point MaxScale starts to the
bitter end, but in a subsequent commit it will be changes so that the
QPS is retained for a sliding window.