MXS-4791 Add SortKey and QueryKey
SortKey estblishes an absolut order of queries by event_id.

Similar to SortKey, QueryKey allows containers of QueryKeys
to be sorted and copied fast.

The QueryKey contains a std::unique_ptr<QueryEvent>.
This could be an iterator to something that holds
QueryEvents, but since that container would have to
have a fast erase that does not invalidate iterators,
a seprate container approach, apart from making the code
more complex, would not add a performance advantage.