MXS-4791 Implement merge-sort
With fake files. class ExternalChunks needs to be
fully implemented.

The algorithm overwrites the input file.
As long as the chunk size is large enough and events are
not too far apart (event 1 starts before 2, but 2 ends before 1),
the events are written directly to the output.
The working chunk is like a sliding window in an effort to never
split into files to be merged later. The worst case is when the
first event (earliest start time) is written last in capture.
In that case the algorithm becomes a regular merge-sort.

It is assumed that however many files have been written, they
can all be opened at the same time to perform the final merge
writing events straight to the output file (which is the same
file as the input file).

There are endless opportunities to improve the merge-sort, e.g.,
once the chunk is split for the first time, there is no need
to continue the algo but instead only write (larger) external files.