MXS-5043 Calculate QPS per routing worker
Now all sessions objects in the same routing worker will update
the same QPS object, which means that the memory usage will be
much smaller compared to the case where it is calucated per
session.

DiffRouter has behind a lock a vector of shared pointers to
a vector of entries that contain a target and a shared pointer
to a DiffQps. The first vector is indexed using the routing
worker index, and the entries in that are shared pointers to
allow resizing.

The second vector is simply looped over to find the DiffQps
associated with a particual target for the routing worker in
question. Conceptually, a map would be better, but for now
the number of targets is no more than 2, so the cost would be
more that the gain.