Fix RHEL 7 and RHEL 8 build failures
Neither of the two has a new enough of a compiler and TBB to
successfully use the <execution> header. The previous method of using a
check program is a simpler approach since the compiler version alone
isn't enough.

Moved the defines into the correct file and modified the new code to use
it.