Don't force a .csv file extension to be used
This prevents redirecting the replay output into /dev/null. The
RepBoostStorage class still requires a .rx ending as the conversion code
expects the ending to be present.

Also modified the file existence check to account for files that exist
but are empty and for things that exist but do not have a size, for
example FIFOs.