MXS-4791 Rename Storage::size() to Storage::tell()
Storage is a file-like object and tell() will be used (or is
already used) to tell the approximate size of bytes read
or written.

Just as a reminder: The base class Storage is not really needed
anymore as CapBoostStorage is the only remaining derived type
and is used directly to access its member functions.
Then again, as a storage class could be written in the
future with custom serialization it does not hurt to keep
class Storage around.