MXS-4761 Rename commands prepare/unprepare to create/destroy
Further, the name of the Diff service to be created must now be
specified explicitly. Earlier the diff service was prepared/created
for the service MyService as.

  maxctrl call command diff prepare MyService Server1 NewServer

and as a result the Diff service DiffMyService was created. Now the
command is

  maxctrl call command diff create MyDiffer MyService Server1 NewServer

As the name must be specified explicitly a possible name-clash is the
problem of the user.

Of symmetry reasons, the command unprepare was renamed to destroy.