Add deadlock monitor thread
The thread detects if a session has waited for 75% of the
innodb_lock_wait_timeout and then proceeds to print the output of SHOW
ENGINE INNODB STATUS which shows database lock ownership. The MariaDB
thread IDs, replay session IDs and the last executed event of each
session is also collected. This helps determine which transactions are
deadlocked.