Allow non-modifying transactions to run at any time
If the simulation time has advanced to the point in time where the
original event was executed, all events that are not part of a
transaction that committed a GTID are allowed to execute.

Previously, the events were only allowed to execute after all the
transactions that committed before it during capture were committed in
replay. This caused frequent deadlock situations if a transactions that
was rolled back and a transaction that committed both took row level
locks but the order in which they were granted was inverted in
replay. By allowing the rolled back transaction to proceed, the deadlock
is solved when the ROLLBACK is executed.

One case that is not handled is when a transaction sees a deadlock error
but proceeds to continue doing writes and later commits it. The implicit
rollback that the database does is not encoded in the capture and doing
so would make sure that the same actions will be done in the replay.