Limit SQL length in visualization
Long batch INSERT statements ended up filling up the screen. The SQL is
now limited to 400 characters and 150 characters in the dropdown menu.

The change in errors also now always includes the sign. The table output
was confusing as the `Errors` column was the change in errors and not
the absolute value of errors. This should hopefully make it clearer.