Add --commit-order option
The option allows transactions to be executed in three possible
modes.

--commit-order=none removes all ordering of transactions and makes it so
that they can be executed as soon as the simulation has advanced to that
point.

--commit-order=optimistic is the current implementation where the
transactions are allowed to start if they started before the current
transaction's end. This means that a transaction that started before the
latest committed transaction than then later committed after it can
start execution. This schedules the events in close to original order
but it still allows out-of-order execution.

--commit-order=serialized causes transactions to be executed in the
order they were committed. This effectively serializes the execution of
all transactions that committed a GTID and removes all non-read-only
parallelism.