Use SHOW ENGINE INNODB STATUS in deadlock monitoring
Since the intended behavior of the deadlock monitor is to detect lock
wait timeouts, it should only report those connections that indeed are
waiting on a lock rather than just processing data. Large analytics and
reporting queries can exceed the lock wait timeout which is why the
check needs to be a bit smarter.

If needed, the thread can be repurposed to monitor other statistics that
are visible during the replay like global InnoDB page metrics but
currently it's unknown whether this is actually a useful metric to
track.