MXS-5077 Calculate histogram spec using results from main
Earlier the spec was calculated using results from both main
and other. With this change it will be trivial to detect when
a result from other goes outside the 99 percentile of main.

That seems like a better trigger for EXPLAIN than using a
difference in result between main and other, as that leads
to a lot of false positives, since there during normal
operation is a lot of variance that is not actually a sign
of a problem.