Convert logical errors into debug assertions
Some of the checks for the `show` commands were duplicated in the
RepShow class that were already done at the upper level which means that
they can be converted into debug assertions.

The code that handled unhandled expections isn't that valuable given
that it only logs a warning when not handling it all should (ideally)
give a better stacktrace if an unhandled exception is thrown.

With these changes, the test coverage of maxplayer is above 90%.