Add WCAR live capture test
The test checks that starting a capture at runtime works as
expected. The test runs both the size-limited and duration-limited
versions of the same workload to also stress those parts of the filter.

The replays are also done in parallel to both speed up the testing as
well as to see that there are no issues in concurrent replays.