Match captured ROLLBACK behavior in replay
If the original transaction ended up being rolled back automatically
with an error 1213, the replay may not necessarily receive the same
error. If the query that in the capture received an error 1213 as the
response does not receive the same error in the replay, the ROLLBACK
must be done "manually" by the session.