MXS-4791 Move state mutex to all public session entry points
The mutex only protected the state change, but later changes
have caused more shared data. Protecting that data and possible
future changes is complex and error prone.

The mutex is per session and only protects for commands coming
from the filter, which essentially are maxctrl user commands or
timeouts thus very rare. There is practically no contention on
the mutex, so the cost is very low. This was also verified by
benchmarking.