MXS-5111: Support new protocol and LDLI in smartrouter
Taking into use the functionality that's exposed by mxs::Reply, the need
for the manual parsing of results is removed. This removes the need for
RCAP_TYPE_OLD_PROTOCOL. Using mariadb::MultiPartTracker and the
functions from the MariaDB protocol removes the need for the
mysql_plus.hh header. Since the only user of that header was the
smartrouter and it did not support the newer protocol extensions, the
header can be removed.