MXS-5111: Remove RCAP_TYPE_OLD_PROTOCOL
Now that all of the modules support all of the protocol extensions, the
capability can be removed. Commenting out the value will help with any
accidental use of it which might result from a merge or a pull request
but it still allows it to be referenced by searching it.