MXS-5105: Always free nodes in vacate_lru()
This is the first step in storing the Nodes by value inside the
mxb::lru_cache. The goal of this is to avoid the extra indirection that
takes place currently: the std::unordered_map stores an iterator
(pointer) to the std::list that itself stores the Node pointer. This is
one extra pointer compared to the old version where the pointer to the
Node was stored in the std::unordered_map.