Check MainWorker before using it in signal handlers
The SIGTERM handler may be called after the MainWorker has stopped if
something causes MaxScale to stop on its own after daemonizing. It also
may happen if MaxScale is very slow to stop which seems to happen when a
GCOV build is being done as the profiling data is written to files on
disk.