MXS-4761 Remove failing test
The assumption that the creation of a Diff service would fail
was incorrect as by default Diff will attempt to reset the
replication at the end.

The assumption seemed to be correct since the monitor usually
had not had time to update the server state. With the addition
of wait_for_monitor() it obviously always has, which causes
the creation attempt to always succeed and hence the test to
fail.