Fix replay of captures with pings or connection resets
The asynchronous reading of the results was done in the wrong place: the
read should've been done in the same `else` branch where the query was
sent. If a ping or a reset connection was sent, the replay would hang
right after it.

The reading of the --analyze result was also always done when it should
only be done for actual SQL commands.