MXS-4761 Fix leaks in Diff
The result object of the 'main' backend and the corresponding result
object of the 'other' backend refer to each other using shared_ptr:s.
That's by design, to ensure that they stay alive until a response has
been received from both backends. When a result object is ready, it
informs its peer, which then will break the circular dependency if it
itself at that point also is ready. If it is not, the circular dependency
will be broken when it becomes ready.
It must be ensured that the circular dependency is broken, also in the
presence of errors that prevent normal processing.